+++ src/components/image-gesture-voice/image-gesture-voice-interface.ts
@@ -21,5 +21,6 @@ export interface IGVData {
 }
 
 export interface IGVOptions {
-  debug: boolean
+  debug?: boolean
+  colors?: string[]
 }